refactor(General): :recycle: Rename file with suffix constant to plural
