+++ src/core/components/SkTable/SkTable.interfaces.ts
