+++ src/pages/Addresses/components/RequestsByAddress.tsx
@@ -10,14 +10,14 @@ import { LinkCellProps } from '@core/components/LinkCell/LinkCell.interfaces';
 import SkTable from '@core/components/SkTable';
 import ViewDetailCell from '@core/components/ViewDetailsCell';
 import { getDataFromSession, storeDataToSession } from '@core/utils/persistData';
-import { ProcessesComponentsTable, processesTableColumns } from '@pages/Processes/Processes.constant';
+import { ProcessesComponentsTable, processesTableColumns } from '@pages/Processes/Processes.constants';
 import FlowsPair from '@pages/shared/FlowPairs/FlowPair';
-import { flowPairsComponentsTable } from '@pages/shared/FlowPairs/FlowPairs.constant';
+import { flowPairsComponentsTable } from '@pages/shared/FlowPairs/FlowPairs.constants';
 import Metrics from '@pages/shared/Metrics';
 import { SelectedFilters } from '@pages/shared/Metrics/Metrics.interfaces';
 import { FlowPairsResponse, RequestOptions } from 'API/REST.interfaces';
 
-import { TAB_0_KEY, TAB_1_KEY, TAB_2_KEY, httpColumns } from '../Addresses.constant';
+import { TAB_0_KEY, TAB_1_KEY, TAB_2_KEY, httpColumns } from '../Addresses.constants';
 import { RequestLabels, AddressesLabels } from '../Addresses.enum';
 import { RequestsByAddressProps } from '../Addresses.interfaces';
 import { QueriesServices } from '../services/services.enum';
