+++ src/pages/Addresses/views/Addresses.tsx
@@ -10,7 +10,7 @@ import { getTestsIds } from '@config/testIds.config';
 import SkTable from '@core/components/SkTable';
 import MainContainer from '@layout/MainContainer';
 
-import { addressesColumns, addressesColumnsWithFlowPairsCounters, customAddressCells } from '../Addresses.constant';
+import { addressesColumns, addressesColumnsWithFlowPairsCounters, customAddressCells } from '../Addresses.constants';
 import { AddressesLabels } from '../Addresses.enum';
 import { AddressesController } from '../services';
 import { QueriesServices } from '../services/services.enum';
